requirements.txt
asgiref==3.8.1
Django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-filter>=23.5,<24.0
pillow==11.0.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
sqlparse==0.5.1
channels>=4.0.0,<4.1.0
python-dotenv>=1.0.0,<1.1.0
Faker>=20.1.0,<20.2.0
pytz>=2023.3,<2024.0
